refactor: refactor and move some classes
