--- src/main/kotlin/com/github/kuro46/embedscript/script/processor/AbstractParser.kt
@@ -1,11 +1,0 @@
-package com.github.kuro46.embedscript.script.processor
-
-abstract class AbstractParser : Processor.Parser {
-    override fun prepareBuild(processor: ScriptProcessor, script: MutableScript, key: String, matchedValues: List<String>) {
-
-    }
-
-    override fun build(builder: ScriptBuilder, key: String, matchedValues: List<String>) {
-        builder.script.putAll(key, matchedValues)
-    }
-}
