+++ src/main/kotlin/com/github/kuro46/embedscript/script/processor/Processors.kt
@@ -2,6 +2,10 @@ package com.github.kuro46.embedscript.script.processor
 
 import com.github.kuro46.embedscript.script.ParseException
 import com.github.kuro46.embedscript.script.Script
+import com.github.kuro46.embedscript.script.processor.executor.AbstractExecutor
+import com.github.kuro46.embedscript.script.processor.executor.Executor
+import com.github.kuro46.embedscript.script.processor.parser.AbstractParser
+import com.github.kuro46.embedscript.script.processor.parser.Parser
 import net.md_5.bungee.chat.ComponentSerializer
 import org.bukkit.Bukkit
 import org.bukkit.entity.Player
@@ -9,10 +13,10 @@ import java.util.Locale
 import java.util.stream.Collectors
 
 object Processors {
-    val DEFAULT_EXECUTOR: Processor.Executor = object : AbstractExecutor() {
+    val DEFAULT_EXECUTOR: Executor = object : AbstractExecutor() {
 
     }
-    val DEFAULT_PARSER: Processor.Parser = object : AbstractParser() {
+    val DEFAULT_PARSER: Parser = object : AbstractParser() {
 
     }
 
@@ -136,8 +140,8 @@ object Processors {
 
     fun newProcessor(key: String,
                      omittedKey: String,
-                     parser: Processor.Parser,
-                     executor: Processor.Executor): Processor {
+                     parser: Parser,
+                     executor: Executor): Processor {
         return object : Processor {
             override val key = key
