+++ src/main/kotlin/com/github/kuro46/embedscript/script/processor/executor/AbstractExecutor.kt
@@ -1,8 +1,8 @@
-package com.github.kuro46.embedscript.script.processor
+package com.github.kuro46.embedscript.script.processor.executor
 
 import org.bukkit.entity.Player
 
-abstract class AbstractExecutor : Processor.Executor {
+abstract class AbstractExecutor : Executor {
     override fun check(trigger: Player, matchedValues: List<String>): Boolean {
         return true
     }
